// src/types/partials/AuditEventAgent.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TReference } from '../partials/Reference';
import { TUri } from '../simpleTypes/Uri';
import { TCoding } from '../partials/Coding';
import { TAuditEventNetwork } from '../partials/AuditEventNetwork';

export type TAuditEventAgent = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    type?: TCodeableConcept;
    role?: TCodeableConcept[];
    who?: TReference;
    altId?: String;
    name?: String;
    requestor: Boolean;
    location?: TReference;
    policy?: TUri[];
    media?: TCoding;
    network?: TAuditEventNetwork;
    purposeOfUse?: TCodeableConcept[];
};